requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
cheroot==8.6.0
click==8.1.3
colorama==0.4.5
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
h11==0.14.0
idna==3.4
jaraco.functools==3.5.1
more-itertools==8.14.0
Pillow==9.2.0
psutil==5.9.2
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
serpent==1.41
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
uvicorn==0.18.3
web.py==0.62
